Implement `fetch_order_credentials` and `prepare_credentials_presentation`
Also breaks SkuSdkFetcher out to its own file

Implementation for work cherry-picked from https://github.com/brave/brave-core/pull/10713
