+++ components/skus/browser/brave-rewards-cxx/src/shim.h
@@ -1,7 +1,17 @@
+// Copyright (c) 2021 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+#ifndef BRAVE_COMPONENTS_SKUS_BROWSER_BRAVE_REWARDS_CXX_SRC_SHIM_H_
+#define BRAVE_COMPONENTS_SKUS_BROWSER_BRAVE_REWARDS_CXX_SRC_SHIM_H_
+
 #include <functional>
 #include <memory>
-#include "cxx.h"
 
+// NOTE: when running cxxbridge or examples/main.cc, comment out the following
+// two lines and instead `#include "cxx.h"` (please don't commit that change).
+#include "brave/components/skus/browser/brave-rewards-cxx/src/cxx.h"
 #include "brave/components/skus/common/skus_sdk.mojom.h"
 
 namespace brave_rewards {
@@ -36,11 +46,13 @@ class PrepareCredentialsPresentationCallbackState {
 using RefreshOrderCallback = void (*)(RefreshOrderCallbackState* callback_state,
                                       RewardsResult result,
                                       rust::cxxbridge1::Str order);
-using FetchOrderCredentialsCallback = void (*)(FetchOrderCredentialsCallbackState* callback_state,
-                                      RewardsResult result);
-using PrepareCredentialsPresentationCallback = void (*)(PrepareCredentialsPresentationCallbackState* callback_state,
-                                      RewardsResult result,
-                                      rust::cxxbridge1::Str presentation);
+using FetchOrderCredentialsCallback =
+    void (*)(FetchOrderCredentialsCallbackState* callback_state,
+             RewardsResult result);
+using PrepareCredentialsPresentationCallback =
+    void (*)(PrepareCredentialsPresentationCallbackState* callback_state,
+             RewardsResult result,
+             rust::cxxbridge1::Str presentation);
 
 void shim_purge();
 void shim_set(rust::cxxbridge1::Str key, rust::cxxbridge1::Str value);
@@ -59,3 +71,5 @@ void shim_executeRequest(
              brave_rewards::HttpResponse)> done,
     rust::cxxbridge1::Box<brave_rewards::HttpRoundtripContext> ctx);
 }  // namespace brave_rewards
+
+#endif  // BRAVE_COMPONENTS_SKUS_BROWSER_BRAVE_REWARDS_CXX_SRC_SHIM_H_
